// src/common/constants/env.enum.ts
export enum EEnv {
  PORT = 'PORT',
  DB_TYPE = 'DB_TYPE',
  DB_PORT = 'DB_PORT',
  DB_USERNAME = 'DB_USERNAME',
  DB_HOST = 'DB_HOST',
  DB_PASSWORD = 'DB_PASSWORD',
  DB_NAME = 'DB_NAME',
  JWT_SECRET = 'JWT_SECRET',
  JWT_EXPIRES_IN = 'JWT_EXPIRES_IN',
  TMN_CODE = 'TMN_CODE',
  SECRET_KEY = 'SECRET_KEY',
  PAYMENT_GATEWAY_URL = 'PAYMENT_GATEWAY_URL',
  CALLBACK_URL = 'CALLBACK_URL',
  PAYMENT_GATEWAY_VERSION = 'PAYMENT_GATEWAY_VERSION',
  USER_JID_TTL = 'USER_JID_TTL',
}